feat: 상품 추가 페이지 기본 정보 영역 구현
- 상위 카테고리 선택 시 하위 카테고리 조회 API 실행

+++ delivery-market-admin/src/main/java/com/ddangme/dmadmin/dto/goods/GoodsDTO.java
@@ -1,6 +1,7 @@
 package com.ddangme.dmadmin.dto.goods;
 
 import com.ddangme.dmadmin.dto.AdminDTO;
+import com.ddangme.dmadmin.dto.category.CategoryDTO;
 import com.ddangme.dmadmin.model.constants.SaleStatus;
 import com.ddangme.dmadmin.model.constants.UploadFile;
 import com.ddangme.dmadmin.model.goods.Goods;
@@ -21,7 +22,7 @@ public class GoodsDTO {
     private String summary;
     private Long price;
     private Long salePrice;
-    private Long salePercent;
+    private Integer salePercent;
     private SaleStatus saleStatus;
     private GoodsDetailDTO goodsDetailDTO;
     private Set<GoodsOptionDTO> goodsOptionDTO;
